[FIX] cast into if

+++ ECS/src/systems/EventSystem.cpp
@@ -26,13 +26,13 @@ void EventSystem::RunEvents(Coordinator &coordinator)
 				coordinator.AddComponent<Position>(entity, {std::any_cast<float>(event._data[1]), std::any_cast<float>(event._data[2])});
 				coordinator.AddComponent<Velocity>(entity, {std::any_cast<float>(event._data[3]), std::any_cast<float>(event._data[4])});
 				coordinator.AddComponent<Hitbox>(entity, {std::any_cast<float>(event._data[5]), std::any_cast<float>(event._data[6])});
-				if (std::any_cast<int>(event._data[9]) == PLAYER)
+				if (std::any_cast<float>(event._data[9]) == PLAYER)
 					coordinator.AddComponent<Sprite>(entity, {Graphic::loadTexture("assets/spaceship.png")});
 				if (std::any_cast<int>(event._data[10]) == static_cast<int>(entity))
-					coordinator.AddComponent<Movable>(entity, {NONE});	
-				if (std::any_cast<int>(event._data[9]) == BULLET)
+					coordinator.AddComponent<Movable>(entity, {NONE});
+				if (std::any_cast<float>(event._data[9]) == BULLET)
 					coordinator.AddComponent<Sprite>(entity, {Graphic::loadTexture("assets/bullets.png")});
-				if (std::any_cast<int>(event._data[9]) == ENNEMY)
+				if (std::any_cast<float>(event._data[9]) == ENNEMY)
 					coordinator.AddComponent<Sprite>(entity, {Graphic::loadTexture("assets/carli.png")});
 				break;
 			} case Event::actions::MOVE: {
